+++ src/index.ts
@@ -3,5 +3,6 @@ export * from "./execute";
 export * from "./http";
 export * from "./introspect";
 export * from "./read";
+export * from "./server";
 export * from "./util";
 export * from "./ws";