+++ src/util/LoggerLike.ts
@@ -1,5 +1,5 @@
 export interface LoggerLike {
-  log: (message: string | Error) => void;
-  warn: (message: string | Error) => void;
-  error: (message: string | Error) => void;
+  log: (message: string | Error) => unknown;
+  warn: (message: string | Error) => unknown;
+  error: (message: string | Error) => unknown;
 }