+++ src/ws/GraphQLWebSocketClient.ts
@@ -1,3 +1,4 @@
+// keep specifically targeted imports here fore browser build
 import { AsyncPushIterator } from "../util/AsyncPushIterator";
 import { toError } from "../util/toError";
 import {